Move folder Profile Components
