+++ web/src/pages/Profile.tsx
@@ -1,6 +1,6 @@
 // (c) Tecnologico de Monterrey 2022, rights reserved.
 
-import ProfileInfo from "../components_profile/ProfileInfo";
+import ProfileInfo from "./components/Profile/ProfileInfo";
 
 const Profile = (): JSX.Element => {
   return (
